Remove event listener from highlights after sidebar is closed.
Previously,  used to highlight and throw errors even after the sidebar is closed.

Also, moved sidebar_overlay content_script related stuff into it's own folder.
