+++ src/direct-linking/content_script/rendering.js
@@ -1,6 +1,6 @@
 import * as AllRaven from 'raven-js'
 import { retryUntil } from '../utils'
-import { attachEventListenersToNewHighlights } from '../../sidebar-overlay/interactions'
+import { attachEventListenersToNewHighlights } from 'src/sidebar-overlay/content_script/interactions'
 import { descriptorToRange, markRange } from './annotations'
 
 import styles from './styles.css'
