+++ src/sidebar-overlay/components/CommentBox.js
@@ -77,7 +77,6 @@ class CommentBox extends React.Component {
     }
 
     save = () => {
-        // TODO: Fix tags not getting stored right
         const { commentInput, tags } = this.state
         const { anchor } = this.props
         const strippedComment = commentInput.trim()
