+++ src/sidebar-overlay/content_script/index.js
