+++ src/sidebar-overlay/content_script/styles.css
@@ -1,0 +1,11 @@
+.memex-highlight {
+    background: rgb(62, 185, 149, 0.4);
+    transition: all 0.3s ease-in;
+    cursor: pointer;
+}
+
+.memex-highlight-dark {
+    background: rgb(62, 185, 149);
+    transition: all 0.3s ease-in;
+    cursor: pointer;
+}
